+++ tests/switch/test_five.py
@@ -1,5 +1,5 @@
 
-from nintendo import five
+from nintendo.switch import five
 from anynet import http
 import contextlib
 import pytest